llm-api/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
dataclasses-json==0.6.7
exceptiongroup==1.2.2
fastapi==0.115.4
frozenlist==1.5.0
gigachain==0.2.16
gigachain-community==0.2.16.post3
gigachain-core==0.2.38.post2
gigachain-text-splitters==0.2.4
gigachat==0.1.35
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langfuse==2.53.3
langsmith==0.1.139
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.11
packaging==24.1
propcache==0.2.0
protobuf==5.28.3
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.2
tenacity==8.5.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wrapt==1.16.0
yandex-speechkit==1.5.0
yarl==1.17.1
